feat(account-pages): moved address and order pages into separate dirs
